Render all panes (#69)
* Move About component from NavMenu to App

* Change redux state for current pane, render all panes

* Update tests

* Update changelog
